server: improve error handling a bit and fill logic
